+++ include/lap.hpp
@@ -28,8 +28,8 @@
 
 struct solution {
   int cost;
-  std::vector<int> rowsol;
-  std::vector<int> colsol;
+  std::vector<size_t> rowsol;
+  std::vector<size_t> colsol;
   std::vector<int> u;
   std::vector<int> v;  
 };
@@ -43,7 +43,7 @@ solution lap(const daestruct::sigma_matrix& cost);
  * Solve the integer linear assignment problem using an older (partiall) assignment
  */
 solution delta_lap(const daestruct::sigma_matrix& assigncost, const std::vector<int>& _u, const std::vector<int>& _v, 
-		   const std::vector<int>& _rowsol, const std::vector<int>& _colsol);
+		   const std::vector<size_t>& _rowsol, const std::vector<size_t>& _colsol);
 
 std::ostream& operator<<(std::ostream& o, const solution& s);
